Extracted OctokitClient from SprintStatistics class
