+++ sprint_statistics.rb
@@ -1,9 +1,10 @@
+require_relative 'octokit_client'
 require 'active_support'
 require 'active_support/core_ext'
 
 class SprintStatistics
-  def initialize(access_token, milestone_string = nil)
-    @access_token = access_token
+  def initialize(github_api_token, milestone_string = nil)
+    @github_api_token = github_api_token
     @milestone_string = milestone_string
   end
 
@@ -32,30 +33,7 @@ class SprintStatistics
   end
 
   def client
-    @client ||= begin
-      require 'octokit'
-      Octokit.configure do |c|
-        c.auto_paginate = true
-        c.middleware = Faraday::RackBuilder.new do |builder|
-          if ENV['DEBUG']
-            builder.use(
-              Class.new(Faraday::Response::Middleware) do
-                def on_complete(env)
-                  api_calls_remaining = env.response_headers['x-ratelimit-remaining']
-                  STDOUT.puts "DEBUG: Executed #{env.method.to_s.upcase} #{env.url} ... api calls remaining #{api_calls_remaining}"
-                end
-              end
-            )
-          end
-
-          builder.use Octokit::Response::RaiseError
-          builder.use Octokit::Response::FeedParser
-          builder.adapter Faraday.default_adapter
-        end
-      end
-
-      Octokit::Client.new(:access_token => @access_token)
-    end
+    @client ||= OctokitClient.new(@github_api_token).handle
   end
 
   def fetch(collection, *args)
